+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Give icons in Gutenberg block admin UI now display correctly in Firefox (#6023)
+- An error message that was confusing has been clarified (#6045)
 
 ## 2.15.0 - 2021-10-11