+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+## 2.16.0 - 2021-10-26
+
 ## 2.16.0-rc.1 - 2021-10-25
 
 ### New